pyproject.toml
[project]
name = "pyrecon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "pywebview>=5.4",
]
